Delete IPTB-based sync test
