+++ functional-tests/mining_chain_test.go
@@ -7,7 +7,10 @@ import (
 	"testing"
 	"time"
 
+	"github.com/filecoin-project/go-filecoin/internal/app/go-filecoin/node/test"
+	"github.com/filecoin-project/go-filecoin/internal/pkg/consensus"
 	"github.com/filecoin-project/go-filecoin/internal/pkg/drand"
+	"github.com/filecoin-project/go-filecoin/internal/pkg/proofs"
 
 	"github.com/filecoin-project/specs-actors/actors/abi"
 	"github.com/stretchr/testify/require"
@@ -135,4 +138,75 @@ func TestSyncFromSingleMiner(t *testing.T) {
 	require.NoError(t, th.WaitForIt(50, 20*time.Millisecond, func() (bool, error) {
 		return ndValidator.Chain().ChainReader.GetHead().Equals(head), nil
 	}), "validator failed to sync new head")
+}
+
+func TestBootstrapWindowedPoSt(t *testing.T) {
+	tf.FunctionalTest(t)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	wd, _ := os.Getwd()
+	genCfgPath := filepath.Join(wd, "..", "fixtures/setup.json")
+	presealPath := filepath.Join(wd, "..", "fixtures/genesis-sectors")
+	// setup presealed sectors and uncomment to run test against sectors with larger sector size
+	//genCfgPath := filepath.Join("./512", "setup.json")
+	//presealPath := "./512"
+
+	genTime := int64(1000000000)
+	blockTime := 30 * time.Second
+	fakeClock := clock.NewFake(time.Unix(genTime, 0))
+
+	// Load genesis config fixture.
+	genCfg := loadGenesisConfig(t, genCfgPath)
+	// set proving period start to something soon
+	start := abi.ChainEpoch(0)
+	genCfg.Miners[0].ProvingPeriodStart = &start
+
+	seed := node.MakeChainSeed(t, genCfg)
+
+	// fake proofs so we can run through a proving period quickly
+	miner := test.NewNodeBuilder(t).
+		WithBuilderOpt(node.ChainClockConfigOption(clock.NewChainClockFromClock(uint64(genTime), blockTime, fakeClock))).
+		WithGenesisInit(seed.GenesisInitFunc).
+		WithBuilderOpt(node.DrandConfigOption(&drand.Fake{
+			GenesisTime:   time.Unix(genTime, 0).Add(-1 * blockTime),
+			FirstFilecoin: 0,
+		})).
+		WithBuilderOpt(node.VerifierConfigOption(&proofs.FakeVerifier{})).
+		WithBuilderOpt(node.PoStGeneratorOption(&consensus.TestElectionPoster{})).
+		Build(ctx)
+
+	_, _, err := initNodeGenesisMiner(ctx, t, miner, seed, genCfg.Miners[0].Owner, presealPath)
+	require.NoError(t, err)
+
+	err = miner.Start(ctx)
+	require.NoError(t, err)
+
+	err = miner.StorageMining.Start(ctx)
+	require.NoError(t, err)
+
+	// mine once to enter proving period
+	go simulateBlockMining(ctx, t, fakeClock, blockTime, miner)
+
+	minerAddr := miner.Repo.Config().Mining.MinerAddress
+
+	// Post should have been triggered, simulate mining while waiting for update to proving period start
+	for i := 0; i < 50; i++ {
+		head := miner.Chain().ChainReader.GetHead()
+
+		view, err := miner.Chain().State.StateView(head)
+		require.NoError(t, err)
+
+		poSts, err := view.MinerSuccessfulPoSts(ctx, minerAddr)
+		require.NoError(t, err)
+
+		if poSts > 0 {
+			return
+		}
+
+		// We need to mine enough blocks to get to get to the deadline that contains our sectors. Add some friction here.
+		time.Sleep(2 * time.Second)
+	}
+	t.Fatal("Timouut waiting for windowed PoSt")
 }
